fix: move dictionaries to constants.js
